Value: stick to {f,u}64 for now
Using u128 is causing issues in the MacOS CI which do not happen on u64.
The current implementation is not ideal anyways as it doesn't support
arbitrary sized integers which are coming in C2X but it is still better
than what we have in master.

+++ src/Diagnostics.zig
@@ -31,8 +31,8 @@ pub const Message = struct {
             resembles: u21,
         },
         actual_codepoint: u21,
-        unsigned: u128,
-        signed: i128,
+        unsigned: u64,
+        signed: i64,
         none: void,
     };
 };
